Fix MSVC compilation
* Make PkgConfig optional

This isn't really common on Windows, so make it optional and only
install PkgConfig files if it's found.

* Add .dll.a to the search path for gcrypt

This is used in some builds of gcrypt on Windows

* Define ssize_t on Windows where gcrypt isn't explicit used

ssize_t is a POSIX'ism (that really should be in Win32, but I digress.)

This is set to long which is the same underlying basic type gcrypt uses
on Windows.